No to ADs, no to Google's especially
Yea you can add an override to the deny list for this, and everything else on here you don't want like the search engine crawlers etc. But for anyone using this, with the whole point being suitable for fedi instances, well if you're one for ads, my fork isn't for you.