Improve the find-tor-exits script
`grep` often seemed to run before `sed` had completed a lot of the time, so it gave off the impression that `sed` was keep missing a bunch of lines constantly and had to be reran which isn't that efficient. However that was just `grep` running before `sed` completed, so letting `sed` do the whole loop first and THEN running `grep` afterwards stops this behavior.