Change script finding tor exits to grab directly from torproject
I was originally getting the list via SecOps-Institute/Tor-IP-Addresses, but I just noticed that's not been updated for over a year lol. So updated to grab the list directly from check.torproject.org and am keeping the list in the repo so I can easily see changes to it to know I'm not using an outdated list.