Rework compile script to escape dots in regex for user agents as well
Upstream escaped dots for referrers but not for user agents. It's till regex and without escaping dots a user agent block such as "fir.fox" would block also "FireFox" aka blocking FireFox users because dot means any character.. so you need to escape it to have the correct behaviour. Thanks to Vampire for alerting about this.

This also tidies up no longer having a separate file specifically for generating referrer regex, that it's all just now the same procedure doing it in the one compile script for all the referrers and user agents. Though be it does add to the amount each section has compared to before, but at least it's proper behaviour now.